# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateUserRequest:
    properties:
      email:
        type: string
      phone:
        type: string
    type: object
...
  handlers.PairDeviceRequest:
    properties:
      device_id:
        type: string
      is_primary:
        type: boolean
      nickname:
        type: string
    type: object
...
  handlers.StatusResponse:
    properties:
      status:
        example: deleted
        type: string
    type: object
...
  handlers.UpdateUserDeviceRequest:
    properties:
      is_primary:
        type: boolean
      nickname:
        type: string
      unpaired_at:
        description: RFC3339 or null
        type: string
    type: object
  handlers.UpdateUserRequest:
    properties:
      email:
        type: string
      phone:
        type: string
    type: object
...
  httpx.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  store.UserDeviceRow:
    properties:
      device_id:
        type: string
      id:
        type: string
      is_primary:
        type: boolean
      nickname:
        type: string
      paired_at:
        type: string
      unpaired_at:
        type: string
      user_id:
        type: string
    type: object
  store.UserRow:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  store.UserWearableRangeDevice:
    properties:
      device_id:
        example: 66a000000000000000000001
        type: string
      deviceUid:
        example: qring-aa-bb-cc-01
        type: string
      is_primary:
        example: true
        type: boolean
      items:
        items:
          $ref: '#/definitions/store.WearableRangeItem'
        type: array
      nickname:
        example: daily ring
        type: string
      paired_at:
        example: "2026-07-01T00:00:00Z"
        type: string
      unpaired_at:
        example: "2026-07-28T00:00:00Z"
        type: string
      user_device_id:
        example: 66a000000000000000000200
        type: string
    type: object
  store.UserWearableRangeResponse:
    properties:
      devices:
        items:
          $ref: '#/definitions/store.UserWearableRangeDevice'
        type: array
      from:
        example: "2026-07-01"
        type: string
      to:
        example: "2026-07-27"
        type: string
      user_id:
        example: 66a000000000000000000100
        type: string
    type: object
  store.WearableActivityReading:
    properties:
      activeTime:
        example: 15
        type: number
      calories:
        example: 8
        type: number
      distance:
        example: 95
        type: number
      steps:
        example: 120
        type: number
      ts:
        example: "2026-07-27T08:00:00.000Z"
        type: string
    type: object
  store.WearableActivitySection:
    properties:
      activeTime:
        example: 29
        type: number
      calories:
        example: 28390
        type: number
      distance:
        example: 493
        type: number
      lastUpdated:
        example: "2026-07-27T23:44:10.512Z"
        type: string
      readings:
        items:
          $ref: '#/definitions/store.WearableActivityReading'
        type: array
      steps:
        example: 717
        type: number
    type: object
  store.WearableBloodPressureReading:
    properties:
      diastolic:
        example: 76
        type: number
      heartRate:
        example: 72
        type: number
      systolic:
        example: 118
        type: number
      ts:
        example: "2026-07-27T07:12:00.000Z"
        type: string
    type: object
  store.WearableBloodPressureSection:
    properties:
      diastolic:
        example: 76
        type: number
      heartRate:
        example: 72
        type: number
      lastUpdated:
        example: "2026-07-27T23:44:10.512Z"
        type: string
      measurementTime:
        example: "2026-07-27T07:12:00.000Z"
        type: string
      readings:
        items:
          $ref: '#/definitions/store.WearableBloodPressureReading'
        type: array
      systolic:
        example: 118
        type: number
    type: object
  store.WearableMetricReading:
    properties:
      raw:
        example: 36.72
        type: number
      ts:
        example: "2026-07-27T23:44:10.512Z"
        type: string
      value:
        example: 77
        type: number
    type: object
  store.WearableMetricSection:
    properties:
      average:
        example: 74
        type: number
      current:
        example: 77
        type: number
      lastUpdated:
        example: "2026-07-27T23:44:10.512Z"
        type: string
      max:
        example: 112
        type: number
      min:
        example: 58
        type: number
      raw:
        example: 36.72
        type: number
      readings:
        items:
          $ref: '#/definitions/store.WearableMetricReading'
        type: array
      readingsCount:
        example: 96
        type: number
    type: object
  store.WearableRangeItem:
    properties:
      activity:
        $ref: '#/definitions/store.WearableActivitySection'
      bloodPressure:
        $ref: '#/definitions/store.WearableBloodPressureSection'
      date:
        example: "2026-07-27"
        type: string
      deviceUid:
        example: qring-aa-bb-cc-01
        type: string
      heartRate:
        $ref: '#/definitions/store.WearableMetricSection'
      hrv:
        $ref: '#/definitions/store.WearableMetricSection'
      sleep:
        $ref: '#/definitions/store.WearableSleepSection'
      spo2:
        $ref: '#/definitions/store.WearableMetricSection'
      stress:
        $ref: '#/definitions/store.WearableMetricSection'
      temperature:
        $ref: '#/definitions/store.WearableMetricSection'
      totalActivities:
        items:
          $ref: '#/definitions/store.WearableTotalActivitySection'
        type: array
      tzOffsetMin:
        example: 420
        type: integer
    type: object
  store.WearableRangeResponse:
    properties:
      deviceUid:
        example: qring-aa-bb-cc-01
        type: string
      from:
        example: "2026-07-01"
        type: string
      items:
        items:
          $ref: '#/definitions/store.WearableRangeItem'
        type: array
      to:
        example: "2026-07-27"
        type: string
    type: object
  store.WearableSleepSection:
    properties:
      awakeMinutes:
        example: 10
        type: number
      deepMinutes:
        example: 92
        type: number
      lastUpdated:
        example: "2026-07-27T23:44:10.512Z"
        type: string
      lightMinutes:
        example: 268
        type: number
      remMinutes:
        example: 61
        type: number
      score:
        example: 78
        type: number
      segments:
        items:
          $ref: '#/definitions/store.WearableSleepSegment'
        type: array
      sleepEnd:
        example: "2026-07-27T05:52:00.000Z"
        type: string
      sleepStart:
        example: "2026-07-26T22:41:00.000Z"
        type: string
      stageData:
        items:
          type: integer
        type: array
      totalMinutes:
        example: 431
        type: number
    type: object
  store.WearableSleepSegment:
    properties:
      end:
        example: "2026-07-26T23:11:00.000Z"
        type: string
      stage:
        example: light
        type: string
      start:
        example: "2026-07-26T22:41:00.000Z"
        type: string
    type: object
  store.WearableSyncRequest:
    properties:
      activity:
        $ref: '#/definitions/store.WearableActivitySection'
      bloodPressure:
        $ref: '#/definitions/store.WearableBloodPressureSection'
      device_uid:
        type: string
      deviceUid:
        type: string
      heartRate:
        $ref: '#/definitions/store.WearableMetricSection'
      hrv:
        $ref: '#/definitions/store.WearableMetricSection'
      sleep:
        $ref: '#/definitions/store.WearableSleepSection'
      spo2:
        $ref: '#/definitions/store.WearableMetricSection'
      stress:
        $ref: '#/definitions/store.WearableMetricSection'
      temperature:
        $ref: '#/definitions/store.WearableMetricSection'
      totalActivities:
        items:
          $ref: '#/definitions/store.WearableTotalActivitySection'
        type: array
      tzOffsetMin:
        type: integer
    type: object
  store.WearableSyncResponse:
    properties:
      device_id:
        example: 66a000000000000000000001
        type: string
      device_uid:
        example: qring-aa-bb-cc-01
        type: string
      status:
        example: synced
        type: string
      sync_date:
        example: "2026-07-27"
        type: string
      upserted:
        $ref: '#/definitions/store.WearableSyncUpsertCounters'
    type: object
  store.WearableSyncUpsertCounters:
    properties:
      daily_activity:
        example: 1
        type: integer
      health_metrics:
        example: 6
        type: integer
      sleep_segments:
        example: 1
        type: integer
      sleep_summary:
        example: 1
        type: integer
      snapshot:
        example: 1
        type: integer
      steps_15m:
        example: 1
        type: integer
      total_activities:
        example: 3
        type: integer
    type: object
  store.WearableTotalActivitySection:
    properties:
      kind:
        example: steps
        type: string
      target:
        example: 4000
        type: number
      value:
        example: 717
        type: number
    type: object
...
info:
  contact: {}
  description: Backend API for AGEM wearable users, devices, qring sensor data, sleep,
    and workouts.
  title: AGEM Backend API
  version: "1.0"
paths:
...
  /v1/user_devices/{id}:
    delete:
      description: Deletes a user-device pairing record.
      parameters:
      - description: Pairing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Delete pairing
      tags:
      - Pairings
    patch:
      consumes:
      - application/json
      description: Updates pairing nickname, primary flag, or unpaired time.
      parameters:
      - description: Pairing ID
        in: path
        name: id
        required: true
        type: string
      - description: Pairing update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.UserDeviceRow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Update pairing
      tags:
      - Pairings
  /v1/users/:
    post:
      consumes:
      - application/json
      description: Creates an AGEM user identity record.
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.UserRow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Create user
      tags:
      - Users
  /v1/users/{id}:
    delete:
      description: Deletes a user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Delete user
      tags:
      - Users
    get:
      description: Returns one AGEM user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.UserRow'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Get user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Updates user email or phone.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.UserRow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Update user
      tags:
      - Users
  /v1/users/{user_id}/devices:
    get:
      description: Lists device pairings for one user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.UserDeviceRow'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List user devices
      tags:
      - Pairings
    post:
      consumes:
      - application/json
      description: Creates a pairing record between a user and a device.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Pairing payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PairDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.UserDeviceRow'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Pair device to user
      tags:
      - Pairings
  /v3/users/{user_id}/wearable/range:
    get:
      description: Returns QRing-style daily wearable snapshots grouped by devices
        paired to the user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Inclusive start date
        format: date
        in: query
        name: from
        required: true
        type: string
      - description: Inclusive end date
        format: date
        in: query
        name: to
        required: true
        type: string
      - description: Include raw readings and sleep segments in snapshot payloads
        in: query
        name: include_readings
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.UserWearableRangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List user wearable snapshots
      tags:
      - Wearable v3
  /v3/wearable/range:
    get:
      description: Returns QRing-style daily wearable snapshots in an inclusive device-local
        date range.
      parameters:
      - description: Stable wearable UID
        in: query
        name: device_uid
        required: true
        type: string
      - description: Alias for device_uid
        in: query
        name: deviceUid
        type: string
      - description: Inclusive start date
        format: date
        in: query
        name: from
        required: true
        type: string
      - description: Inclusive end date
        format: date
        in: query
        name: to
        required: true
        type: string
      - description: Include raw readings and sleep segments in snapshot payloads
        in: query
        name: include_readings
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WearableRangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: List synced wearable snapshots
      tags:
      - Wearable v3
  /v3/wearable/sync:
    post:
      consumes:
      - application/json
      description: Stores a QRing-style daily wearable payload, preserves the snapshot,
        and fans out recognized sections into MongoDB collections.
      parameters:
      - description: Wearable sync payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/store.WearableSyncRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.WearableSyncResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Sync wearable snapshot
      tags:
      - Wearable v3
swagger: "2.0"